cleaned up deckbuilding and renamed to cardBuilder. added some folders to file structure.
