+++ src/scripts/cardBuilder.js
@@ -1,0 +1,18 @@
+import Card from "./cardFoundations/card";
+
+// the truest form of makeCard.
+// takes either no arguments and spits out a blank card, or takes as many templates and applied them to the card
+// example template would be makeCard(Playing('2', '♦'))
+// or => const newPlayingCard = Playing('2', '♦');
+// makeCard(newPlayingCard)
+function makeCard() {
+  const cardBase = Card(); // makes the base of the card
+  const args = [...arguments]; // takes all arguments entered and put them in an array
+  const newCard = Object.assign({}, cardBase, ...args); // returns a new object made from the base card, and any other objects passed
+
+  // return Object.assign({}, Card(), ...arguments); Because I like to shortform things, this line is all thats needed.
+
+  return newCard;
+}
+
+export default makeCard;
